Move NoAbilityPredicate to MuragandaPetroglyphs file
